refactor: MyPostService로 관련 메서드 모아두기
