+++ src/main/java/com/dku/council/domain/like/service/impl/CachedPostLikeServiceImpl.java
@@ -11,16 +11,14 @@ import com.dku.council.domain.post.repository.PostRepository;
 import com.dku.council.domain.user.model.entity.User;
 import com.dku.council.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
-@Primary
 @Service
 @RequiredArgsConstructor
-public class RedisPostLikeServiceImpl implements PostLikeService {
+public class CachedPostLikeServiceImpl implements PostLikeService {
     private final PostLikeMemoryRepository memoryRepository;
     private final PostLikePersistenceRepository persistenceRepository;
     private final UserRepository userRepository;
