+++ src/main/java/com/dku/council/domain/user/service/MyPostService.java
@@ -1,5 +1,10 @@
 package com.dku.council.domain.user.service;
 
+import com.dku.council.domain.comment.CommentRepository;
+import com.dku.council.domain.comment.model.dto.CommentedPostResponseDto;
+import com.dku.council.domain.comment.model.entity.Comment;
+import com.dku.council.domain.like.model.entity.PostLike;
+import com.dku.council.domain.like.repository.PostLikePersistenceRepository;
 import com.dku.council.domain.like.service.PostLikeService;
 import com.dku.council.domain.post.model.dto.list.SummarizedGenericPostDto;
 import com.dku.council.domain.post.model.entity.Post;
@@ -19,6 +24,8 @@ public class MyPostService {
     private final PostRepository postRepository;
     private final FileUploadService fileUploadService;
     private final PostLikeService postLikeService;
+    private final PostLikePersistenceRepository postLikeRepository;
+    private final CommentRepository commentRepository;
 
 
     @Transactional(readOnly = true)
@@ -30,5 +37,25 @@ public class MyPostService {
                         bodySize,
                         postLikeService.getCountOfLikes(e.getId()), e)
         );
+    }
+
+    @Transactional(readOnly = true)
+    public Page<SummarizedGenericPostDto> listMyLikedPosts(Long userId, Pageable pageable, int bodySize){
+        Page<PostLike> list = postLikeRepository.findAllByUserId(userId, pageable);
+
+        return list.map(
+                e -> new SummarizedGenericPostDto(fileUploadService.getBaseURL(),
+                        bodySize,
+                        postLikeService.getCountOfLikes(e.getPost().getId()), e.getPost())
+        );
+    }
+
+    @Transactional(readOnly = true)
+    public Page<CommentedPostResponseDto> listMyCommentedPosts(Long userId, Pageable pageable){
+        Page<Comment> comments = commentRepository.findAllByUserId(userId, pageable);
+        return comments.map(comment -> {
+            Post post = comment.getPost();
+            return new CommentedPostResponseDto(post, comment);
+        });
     }
 }
